+++ frontend/components/features/datasets/datasets-settings/LeftDatasetSettings.content.vue
@@ -29,7 +29,7 @@
         :isLoading="isLoading"
       />
     </div>
-    <div class="delete-dataset-component" v-if="datasetTask">
+    <div class="delete-dataset-component" v-if="datasetTask && !isLoading">
       <DatasetDeleteComponent
         :datasetId="datasetId"
         :datasetTask="datasetTask"