+++ frontend/pages/dataset/_id/annotation-mode/index.vue
@@ -32,7 +32,9 @@
       />
     </template>
     <template v-slot:top>
-      <DatasetFiltersComponent :datasetId="datasetId" />
+      <section class="dataset__top-area">
+        <DatasetFiltersComponent :datasetId="datasetId" />
+      </section>
     </template>
     <template v-slot:center>
       <RecordFeedbackTaskAndQuestionnaireContent :datasetId="datasetId" />
@@ -82,3 +84,13 @@ export default {
   },
 };
 </script>
+
+<style lang="scss" scoped>
+.dataset {
+  &__top-area {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
+</style>