+++ src/argilla/cli/server/database/__main__.py
@@ -18,7 +18,7 @@ from .migrate import migrate_db
 from .revisions import revisions
 from .users import app as users_app
 
-app = typer.Typer(help="Holds CLI commands for migrations and database management.", no_args_is_help=True)
+app = typer.Typer(help="Commands for Argilla server database management", no_args_is_help=True)
 
 app.add_typer(users_app, name="users")
 app.command(name="migrate", help="Run database migrations.")(migrate_db)